+++ packages/subgraph/src/dao/dao.ts
@@ -1,4 +1,4 @@
-import { Address, Bytes, store } from '@graphprotocol/graph-ts';
+import {Address, Bytes, store} from '@graphprotocol/graph-ts';
 
 import {
   DAO as DAOContract,
@@ -8,18 +8,23 @@ import {
   NativeTokenDeposited,
   Granted,
   Frozen,
-  Revoked
+  Revoked,
+  Withdrawn,
+  TrustedForwarderSet,
+  SignatureValidatorSet,
+  StandardCallbackRegistered
 } from '../../generated/templates/DaoTemplate/DAO';
 import {
   Dao,
   ContractPermissionId,
   Permission,
-  VaultTransfer
+  VaultTransfer,
+  StandardCallback
 } from '../../generated/schema';
 
-import { ADDRESS_ZERO } from '../utils/constants';
-import { handleERC20Token, updateBalance } from '../utils/tokens';
-import { addPackage, decodeWithdrawParams, removePackage } from './utils';
+import {ADDRESS_ZERO} from '../utils/constants';
+import {handleERC20Token, updateBalance} from '../utils/tokens';
+import {addPackage, decodeWithdrawParams, removePackage} from './utils';
 
 export function handleMetadataSet(event: MetadataSet): void {
   let daoId = event.address.toHexString();
@@ -66,9 +71,8 @@ export function handleDeposited(event: Deposited): void {
   entity.reference = event.params._reference;
   entity.transaction = event.transaction.hash.toHexString();
   entity.createdAt = event.block.timestamp;
-  entity.type = "Deposit";
+  entity.type = 'Deposit';
   entity.save();
-
 }
 
 export function handleNativeTokenDeposited(event: NativeTokenDeposited): void {
@@ -102,7 +106,7 @@ export function handleNativeTokenDeposited(event: NativeTokenDeposited): void {
   entity.reference = 'Eth deposit';
   entity.transaction = event.transaction.hash.toHexString();
   entity.createdAt = event.block.timestamp;
-  entity.type = "Deposit";
+  entity.type = 'Deposit';
   entity.save();
 }
 
@@ -162,7 +166,13 @@ export function handleExecuted(event: Executed): void {
         '_' +
         event.transactionLogIndex.toHexString() +
         '_' +
-        index.toString();
+        withdrawParams.to +
+        '_' +
+        withdrawParams.amount +
+        '_' +
+        tokenId +
+        '_' +
+        withdrawParams.reference;
 
       let vaultWithdrawEntity = new VaultTransfer(withdrawId);
       vaultWithdrawEntity.dao = daoId;
@@ -173,7 +183,7 @@ export function handleExecuted(event: Executed): void {
       vaultWithdrawEntity.proposal = proposalId;
       vaultWithdrawEntity.transaction = event.transaction.hash.toHexString();
       vaultWithdrawEntity.createdAt = event.block.timestamp;
-      vaultWithdrawEntity.type = "Withdraw";
+      vaultWithdrawEntity.type = 'Withdraw';
       vaultWithdrawEntity.save();
     }
   }
@@ -269,4 +279,87 @@ export function handleFrozen(event: Frozen): void {
   }
   contractPermissionIdEntity.frozen = true;
   contractPermissionIdEntity.save();
+}
+
+export function handleWithdrawn(event: Withdrawn): void {
+  let daoId = event.address.toHexString();
+  let token = event.params.token;
+  let tokenId = handleERC20Token(token);
+  let withdrawnId =
+    daoId +
+    '_' +
+    event.transaction.hash.toHexString() +
+    '_' +
+    event.transactionLogIndex.toHexString() +
+    '_' +
+    event.params.to +
+    '_' +
+    event.params.amount +
+    '_' +
+    tokenId +
+    '_' +
+    event.params._reference;
+  let balanceId = `${daoId}_${token.toHexString()}`;
+
+  let entity = VaultTransfer.load(withdrawnId);
+  if (entity) {
+    return;
+  }
+
+  updateBalance(
+    balanceId,
+    event.address,
+    token,
+    event.params.amount,
+    false,
+    event.block.timestamp
+  );
+
+  entity = new VaultTransfer(withdrawnId);
+  entity.dao = daoId;
+  entity.token = tokenId;
+  entity.sender = event.address;
+  entity.to = event.params.to;
+  entity.amount = event.params.amount;
+  entity.reference = event.params._reference;
+  entity.transaction = event.transaction.hash.toHexString();
+  entity.createdAt = event.block.timestamp;
+  entity.type = 'Withdraw';
+  entity.save();
+}
+
+export function handleTrustedForwarderSet(event: TrustedForwarderSet): void {
+  let daoId = event.address.toHexString();
+  let entity = Dao.load(daoId);
+  if (entity) {
+    entity.trustedForwarder = event.params.forwarder;
+    entity.save();
+  }
+}
+
+export function handleSignatureValidatorSet(
+  event: SignatureValidatorSet
+): void {
+  let daoId = event.address.toHexString();
+  let entity = Dao.load(daoId);
+  if (entity) {
+    entity.signatureValidator = event.params.signatureValidator;
+    entity.save();
+  }
+}
+
+export function handleStandardCallbackRegistered(
+  event: StandardCallbackRegistered
+): void {
+  let daoId = event.address.toHexString();
+  let entityId = `${daoId}_${event.params.interfaceId.toHexString()}`;
+  let entity = StandardCallback.load(entityId);
+  if (!entity) {
+    entity = new StandardCallback(entityId);
+    entity.dao = daoId;
+  }
+  entity.interfaceId = event.params.interfaceId;
+  entity.callbackSelector = event.params.callbackSelector;
+  entity.magicNumber = event.params.magicNumber;
+  entity.save();
 }