+++ packages/subgraph/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adds support for `PluginRepo`, `PluginRegistry` and `PluginSetupProcessor`
+- Added `Withdrawn`, `TrustedForwarderSet`, `StandardCallbackRegistered` and `handleStandardCallbackRegistered` events handling to `DaoTemplate`
 
 ### Changed
 
@@ -29,7 +30,6 @@ On 2022-10-07 15:20:00
 On 2022-10-03 10:38:36
 
 ### Added
-
 - Added `members` to `ERC20VotingPackage`.
 - Added `lastUpdated` to `ERC20VotingVoter`.
 - added `voteCount` to both `ERC20VotingProposal` and `AllowlistProposal`.